backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
opencv-contrib-python